cast double to float

+++ 0x10-variadic_functions/3-print_all.c
@@ -23,7 +23,7 @@ void print_all(const char * const format, ...)
 				printf("%i", va_arg(type, int));
 				break;
 			case 'f':
-				printf("%f", va_arg(type, double));
+				printf("%f",(float) va_arg(type, double));
 				break;
 			case 's':
 				str = va_arg(type, char *);
